chore: move color utils to utils dir
